Combine UserTable and UserList components
